requirements.txt
openpipe-art[skypilot]>=0.1.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
python-dotenv>=1.0.0
torch>=2.0.0
polars>=0.20.0
transformers>=4.0.0
tblib>=3.0.0